+++ CashBault/Model/Models/ErrorModel.swift
