+++ CashBault/Resources/Colors.xcassets/DarkBackground.colorset/Contents.json
@@ -2,7 +2,7 @@
   "colors" : [
     {
       "color" : {
-        "color-space" : "display-p3",
+        "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
           "blue" : "0x25",
