--- CashBault/Views/Cards/CardsView.swift
@@ -1,18 +1,0 @@
-//
-//  CardsView.swift
-//  CashBault
-//
-//  Created by manueltirado on 29/4/24.
-//
-
-import SwiftUI
-
-struct CardsView: View {
-    var body: some View {
-        Text("Cards View")
-    }
-}
-
-#Preview {
-    CardsView()
-}
